Preferencially enable or disable "snoop-precompilation" with `set_snoop_precompile!`
- Enable or disable the snoop-precompilation by passing the appropriate boolean value.
- By default it is enabled aka `true`.
- These preferences are stored in a LocalPrefernces.toml. Instead, to store it in (OrdinaryDiffEq)Project.toml, set `export_prefs = true`
