+++ Project.toml
@@ -27,6 +27,7 @@ NLsolve = "2774e3e8-f4cf-5e23-947b-6d7e65073b56"
 NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
 Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 PreallocationTools = "d236fae5-4411-538c-8e31-a6e3d9e00b46"
+Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
